+++ src/textual/widgets/_directory_tree.py
@@ -75,7 +75,7 @@ class DirectoryTree(TreeControl[DirEntry]):
             await node.add(entry.name, DirEntry(entry.path, entry.is_dir()))
         node.loaded = True
         await node.expand()
-        # self.refresh(layout=True)
+        self.refresh(layout=True)
 
     async def message_tree_click(self, message: TreeClick[DirEntry]) -> None:
         dir_entry = message.node.data